+++ ricloud/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = "3.0.0rc2"
+__version__ = "3.0.0rc3"
 
 from ricloud import conf